Update grmsd.R
Exposed 'method' argument of the 'impute.yai' function to calls of 'grmsd' and/or 'varSelection', as it stands it appears as though every call of impute.yai from these functions only permits the use of the default option ("closest"). Unless mistaken, this would mean that the 'varSelection' function can only help the user find a model appropriate for nearest-neighbour imputation, with all higher values of k being effectively redundant to the variable selection process.